minimax implementation
With the use of AI and a video from YouTube that showed a pseudocode for the minimax function, I was able to get a prototype for mu solution. However the auto grader was not giving the full marks and the error indicated that there was an issue with how the depth was being implemented. Upon further prompting the AI about this issue I was able to get the issue resolved and the auto grader then granted full marks
YouTube link: https://youtu.be/l-hh51ncgDI?feature=shared